// feature-1.14.0/src/types/request.d.ts
export declare namespace RequestParams {
// ...
  interface CreateIteration {
    notify_url: string;
    backend_manager: {
      staff_name: string | null;
      staff_no: string | null;
    };
    dev_time: string | null;
    frontend_manager: {
      staff_name: string | null;
      staff_no: string | null;
    };
    id: number | null;
    iteration_manager: {
      staff_name: string | null;
      staff_no: string | null;
    };
    name: string | null;
    is_schedule: number;
    plan_id: number | null;
    release_time: string | null;
    test_manager: {
      staff_name: string | null;
      staff_no: string | null;
    };
    test_time: string | null;
    union_time: string | null;
  }
// ...
}